requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
paramiko==3.4.0
pywinrm==0.4.3
httpx==0.27.0
apscheduler==3.10.4
openpyxl==3.1.2
scikit-learn==1.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pystray==0.19.5
Pillow==10.3.0
cryptography==42.0.7
python-multipart==0.0.9
